Compression tool upgrades

+++ tools/configurator.sh
@@ -421,28 +421,32 @@ configurator_retroarch_options_dialog() {
 configurator_compress_single_game_dialog() {
   local file=$(file_browse "Game to compress")
   if [[ ! -z "$file" ]]; then
-    if [[ $(validate_for_chd "$file") == "true" ]]; then
+    local compatible_compression_format=$(find_compatible_compression_format "$file")
+    if [[ ! $compatible_compression_format == "none" ]]; then
       local post_compression_cleanup=$(configurator_compression_cleanup_dialog)
-      local filename_no_path=$(basename "$file")
-      local filename_no_extension="${filename_no_path%.*}"
-      local source_file=$(dirname "$(realpath "$file")")"/"$(basename "$file")
-      local dest_file=$(dirname "$(realpath "$file")")"/""$filename_no_extension"
       (
-      echo "# Compressing $filename_no_path, please wait..."
-      compress_to_chd "$source_file" "$dest_file"
-      if [[ $post_compression_cleanup == "true" ]]; then # Remove file(s) if requested
-        if [[ "$file" == *".cue" ]]; then
-          local cue_bin_files=$(grep -o -P "(?<=FILE \").*(?=\".*$)" "$file")
-          local file_path=$(dirname "$(realpath "$file")")
-          while IFS= read -r line
-          do
-            echo "# Removing file $line"
-            rm -f "$file_path/$line"
-          done < <(printf '%s\n' "$cue_bin_files")
-          echo "# Removing file $filename_no_path"
-          rm -f "$file"
-        else
-          echo "# Removing file $filename_no_path"
+      if [[ $compatible_compression_format == "chd" ]]; then
+        if [[ $(validate_for_chd "$file") == "true" ]]; then
+          echo "# Compressing $(basename "$file") to $compatible_compression_format format"
+          compress_game "chd" "$file"
+          if [[ $post_compression_cleanup == "true" ]]; then # Remove file(s) if requested
+            if [[ "$file" == *".cue" ]]; then
+              local cue_bin_files=$(grep -o -P "(?<=FILE \").*(?=\".*$)" "$file")
+              local file_path=$(dirname "$(realpath "$file")")
+              while IFS= read -r line
+              do
+                rm -f "$file_path/$line"
+              done < <(printf '%s\n' "$cue_bin_files")
+              rm -f "$file"
+            else
+              rm -f "$file"
+            fi
+          fi
+        fi
+      else
+        echo "# Compressing $(basename "$file") to $compatible_compression_format format"
+        compress_game "$compatible_compression_format" "$file"
+        if [[ $post_compression_cleanup == "true" ]]; then # Remove file(s) if requested
           rm -f "$file"
         fi
       fi
@@ -452,8 +456,9 @@ configurator_compress_single_game_dialog() {
       --title "RetroDECK Configurator Utility - Compression in Progress"
       configurator_generic_dialog "The compression process is complete!"
       configurator_compress_games_dialog
+
     else
-      configurator_generic_dialog "File type not recognized. Supported file types are .cue, .gdi and .iso"
+      configurator_generic_dialog "The selected file does not have any compatible compressed format."
       configurator_compress_games_dialog
     fi
   else
@@ -462,27 +467,55 @@ configurator_compress_single_game_dialog() {
   fi
 }
 
-configurator_compress_multi_game_dialog() {
-  # This dialog will display any games it finds to be compressable, from the systems listed under each compression type in
-  local compression_format=$1
-  local compressable_game=""
+configurator_compress_some_games_dialog() {
+  # This dialog will display any games it finds to be compressable, from the systems listed under each compression type in compression_targets.cfg
+
   local compressable_games_list=()
   local all_compressable_games=()
-  local compressable_systems_list=$(sed -n '/\['"$compression_format"'\]/, /\[/{ /\['"$compression_format"'\]/! { /\[/! p } }' $compression_targets | sed '/^$/d')
+  local games_to_compress=()
+
+  if [[ ! -z "$1" ]]; then
+    local compression_format="$1"
+  else
+    local compression_format="all"
+  fi
+
+  if [[ $compression_format == "all" ]]; then
+    local compressable_systems_list=$(cat $compression_targets | sed '/^$/d' | sed '/^\[/d')
+  else
+    local compressable_systems_list=$(sed -n '/\['"$compression_format"'\]/, /\[/{ /\['"$compression_format"'\]/! { /\[/! p } }' $compression_targets | sed '/^$/d')
+  fi
 
   while IFS= read -r system # Find and validate all games that are able to be compressed with this compression type
   do
-    if [[ $compression_format == "chd" ]]; then
-      compression_candidates=$(find "$roms_folder/$system" -type f \( -name "*.cue" -o -name "*.iso" -o -name "*.gdi" \) ! -path "*.m3u*")
-    # TODO: Add ZIP file compression search here
+    compression_candidates=$(find "$roms_folder/$system" -type f -not -iname "*.txt")
+    if [[ ! -z $compression_candidates ]]; then
+      while IFS= read -r game
+      do
+        local compatible_compression_format=$(find_compatible_compression_format "$game")
+        if [[ $compression_format == "chd" ]]; then
+          if [[ $compatible_compression_format == "chd" ]]; then
+            all_compressable_games=("${all_compressable_games[@]}" "$game")
+            compressable_games_list=("${compressable_games_list[@]}" "false" "${game#$roms_folder}" "$game")
+          fi
+        elif [[ $compression_format == "zip" ]]; then
+          if [[ $compatible_compression_format == "zip" ]]; then
+            all_compressable_games=("${all_compressable_games[@]}" "$game")
+            compressable_games_list=("${compressable_games_list[@]}" "false" "${game#$roms_folder}" "$game")
+          fi
+        elif [[ $compression_format == "rvz" ]]; then
+          if [[ $compatible_compression_format == "rvz" ]]; then
+            all_compressable_games=("${all_compressable_games[@]}" "$game")
+            compressable_games_list=("${compressable_games_list[@]}" "false" "${game#$roms_folder}" "$game")
+          fi
+        elif [[ $compression_format == "all" ]]; then
+          if [[ ! $compatible_compression_format == "none" ]]; then
+            all_compressable_games=("${all_compressable_games[@]}" "$game")
+            compressable_games_list=("${compressable_games_list[@]}" "false" "${game#$roms_folder}" "$game")
+          fi
+        fi
+      done < <(printf '%s\n' "$compression_candidates")
     fi
-    while IFS= read -r game
-    do
-      if [[ $(validate_for_chd "$game") == "true" ]]; then
-        all_compressable_games=("${all_compressable_games[@]}" "$game")
-        compressable_games_list=("${compressable_games_list[@]}" "false" "${game#$roms_folder}" "$game")
-      fi
-    done < <(printf '%s\n' "$compression_candidates")
   done < <(printf '%s\n' "$compressable_systems_list")
 
   choice=$(zenity \
@@ -497,76 +530,104 @@ configurator_compress_multi_game_dialog() {
 
   local rc=$?
   if [[ $rc == "0" && ! -z $choice ]]; then # User clicked "Compress Selected" with at least one game selected
-    local post_compression_cleanup=$(configurator_compression_cleanup_dialog)
     IFS="," read -ra games_to_compress <<< "$choice"
+    local total_games_to_compress=${#games_to_compress[@]}
+    local games_left_to_compress=$total_games_to_compress
+  elif [[ ! -z $choice ]]; then # User clicked "Compress All"
+    games_to_compress=("${all_compressable_games[@]}")
+    local total_games_to_compress=${#all_compressable_games[@]}
+    local games_left_to_compress=$total_games_to_compress
+  fi
+
+  if [[ ! $(echo "${#games_to_compress[@]}") == "0" ]]; then
+    local post_compression_cleanup=$(configurator_compression_cleanup_dialog)
     (
     for file in "${games_to_compress[@]}"; do
-      local filename_no_path=$(basename "$file")
-      local filename_no_extension="${filename_no_path%.*}"
-      local source_file=$(dirname "$(realpath "$file")")"/"$(basename "$file")
-      local dest_file=$(dirname "$(realpath "$file")")"/""$filename_no_extension"
-      echo "# Compressing $filename_no_path" # Update Zenity dialog text
-      compress_to_chd "$source_file" "$dest_file"
+      local compression_format=$(find_compatible_compression_format "$file")
+      echo "# Compressing $(basename "$file") into $compression_format format" # Update Zenity dialog text
+      progress=$(( 100 - (( 100 / "$total_games_to_compress" ) * "$games_left_to_compress" )))
+      echo $progress
+      games_left_to_compress=$((games_left_to_compress-1))
+      compress_game "$compression_format" "$file"
       if [[ $post_compression_cleanup == "true" ]]; then # Remove file(s) if requested
         if [[ "$file" == *".cue" ]]; then
           local cue_bin_files=$(grep -o -P "(?<=FILE \").*(?=\".*$)" "$file")
           local file_path=$(dirname "$(realpath "$file")")
           while IFS= read -r line
           do
-            echo "# Removing file $line"
             rm -f "$file_path/$line"
           done < <(printf '%s\n' "$cue_bin_files")
-          echo "# Removing file $filename_no_path"
           rm -f $(realpath "$file")
         else
-          echo "# Removing file $filename_no_path"
           rm -f "$(realpath "$file")"
         fi
       fi
     done
     ) |
-    zenity --icon-name=net.retrodeck.retrodeck --progress --no-cancel --pulsate --auto-close \
+    zenity --icon-name=net.retrodeck.retrodeck --progress --no-cancel --auto-close \
       --window-icon="/app/share/icons/hicolor/scalable/apps/net.retrodeck.retrodeck.svg" \
       --title "RetroDECK Configurator Utility - Compression in Progress"
       configurator_generic_dialog "The compression process is complete!"
       configurator_compress_games_dialog
   else
-    if [[ ! -z $choice ]]; then # User clicked "Compress All"
-      local post_compression_cleanup=$(configurator_compression_cleanup_dialog)
-      (
-      for file in "${all_compressable_games[@]}"; do
-        local filename_no_path=$(basename "$file")
-        local filename_no_extension="${filename_no_path%.*}"
-        local source_file=$(dirname "$(realpath "$file")")"/"$(basename "$file")
-        local dest_file=$(dirname "$(realpath "$file")")"/""$filename_no_extension"
-        echo "# Compressing $filename_no_path" # Update Zenity dialog text
-        compress_to_chd "$source_file" "$dest_file"
-        if [[ $post_compression_cleanup == "true" ]]; then # Remove file(s) if requested
-          if [[ "$file" == *".cue" ]]; then
-            local cue_bin_files=$(grep -o -P "(?<=FILE \").*(?=\".*$)" "$file")
-            local file_path=$(dirname "$(realpath "$file")")
-            while IFS= read -r line
-            do
-              echo "# Removing file $line"
-              rm -f "$file_path/$line"
-            done < <(printf '%s\n' "$cue_bin_files")
-            echo "# Removing file $filename_no_path"
-            rm -f $(realpath "$file")
-          else
-            echo "# Removing file $filename_no_path"
-            rm -f $(realpath "$file")
-          fi
+    configurator_compress_games_dialog
+  fi
+}
+
+configurator_compress_all_games_dialog() {
+  # This dialog compress all games found in all compatible roms folders into compatible formats
+
+  local all_compressable_games=()
+  local compressable_systems_list=$(cat $compression_targets | sed '/^$/d' | sed '/^\[/d')
+
+  while IFS= read -r system # Find and validate all games that are able to be compressed with this compression type
+  do
+    compression_candidates=$(find "$roms_folder/$system" -type f -not -iname "*.txt")
+    if [[ ! -z $compression_candidates ]]; then
+      while IFS= read -r game
+      do
+        local compatible_compression_format=$(find_compatible_compression_format "$game")
+        if [[ ! $compatible_compression_format == "none" ]]; then
+          all_compressable_games=("${all_compressable_games[@]}" "$game")
         fi
-      done
+      done < <(printf '%s\n' "$compression_candidates")
+    fi
+  done < <(printf '%s\n' "$compressable_systems_list")
+
+  if [[ ! $(echo ${all_compressable_games[@]}) == "0" ]]; then
+    local post_compression_cleanup=$(configurator_compression_cleanup_dialog)
+    total_games_to_compress=${#all_compressable_games[@]}
+    games_left_to_compress=$total_games_to_compress
+    (
+    for file in "${all_compressable_games[@]}"; do
+      local compression_format=$(find_compatible_compression_format "$file")
+      echo "# Compressing $(basename "$file") into $compression_format format" # Update Zenity dialog text
+      progress=$(( 100 - (( 100 / "$total_games_to_compress" ) * "$games_left_to_compress" )))
+      echo $progress
+      games_left_to_compress=$((games_left_to_compress-1))
+      compress_game "$compression_format" "$file"
+      if [[ $post_compression_cleanup == "true" ]]; then # Remove file(s) if requested
+        if [[ "$file" == *".cue" ]]; then
+          local cue_bin_files=$(grep -o -P "(?<=FILE \").*(?=\".*$)" "$file")
+          local file_path=$(dirname "$(realpath "$file")")
+          while IFS= read -r line
+          do
+            rm -f "$file_path/$line"
+          done < <(printf '%s\n' "$cue_bin_files")
+          rm -f $(realpath "$file")
+        else
+          rm -f $(realpath "$file")
+        fi
+      fi
+    done
     ) |
-    zenity --icon-name=net.retrodeck.retrodeck --progress --no-cancel --pulsate --auto-close \
+    zenity --icon-name=net.retrodeck.retrodeck --progress --no-cancel --auto-close \
     --window-icon="/app/share/icons/hicolor/scalable/apps/net.retrodeck.retrodeck.svg" \
     --title "RetroDECK Configurator Utility - Compression in Progress"
     configurator_generic_dialog "The compression process is complete!"
     configurator_compress_games_dialog
-    else
-      configurator_compress_games_dialog
-    fi
+  else
+    configurator_generic_dialog "There were no games found that could be compressed."
   fi
 }
 
@@ -588,7 +649,11 @@ configurator_compress_games_dialog() {
   --window-icon="/app/share/icons/hicolor/scalable/apps/net.retrodeck.retrodeck.svg" --width=1200 --height=720 \
   --column="Choice" --column="Action" \
   "Compress Single Game" "Compress a single game into a compatible format" \
-  "Compress Multiple Games - CHD" "Compress one or more games compatible with the CHD format" )
+  "Compress Multiple Games - CHD" "Compress one or more games compatible with the CHD format" \
+  "Compress Multiple Games - ZIP" "Compress one or more games compatible with the ZIP format" \
+  "Compress Multiple Games - RVZ" "Compress one or more games compatible with the RVZ format" \
+  "Compress Multiple Games - All Formats" "Compress one or more games compatible with any format" \
+  "Compress All Games" "Compress all games into compatible formats" )
 
   case $choice in
 
@@ -597,10 +662,24 @@ configurator_compress_games_dialog() {
   ;;
 
   "Compress Multiple Games - CHD" )
-    configurator_compress_multi_game_dialog "chd"
+    configurator_compress_some_games_dialog "chd"
   ;;
 
-  # TODO: Add ZIP compression option
+  "Compress Multiple Games - ZIP" )
+    configurator_compress_some_games_dialog "zip"
+  ;;
+
+  "Compress Multiple Games - RVZ" )
+    configurator_compress_some_games_dialog "rvz"
+  ;;
+
+  "Compress Multiple Games - All Formats" )
+    configurator_compress_some_games_dialog "all"
+  ;;
+
+  "Compress All Games" )
+    configurator_compress_all_games_dialog
+  ;;
 
   "" ) # No selection made or Back button clicked
     configurator_welcome_dialog
@@ -620,7 +699,7 @@ configurator_check_multifile_game_structure() {
   else
     configurator_generic_dialog "No incorrect multi-file game folder structures found."
   fi
-  configurator_troubleshooting_tools_dialog
+  configurator_tools_and_troubleshooting_dialog
 }
 
 configurator_check_bios_files_basic() {
@@ -647,7 +726,7 @@ configurator_check_bios_files_basic() {
 
   configurator_generic_dialog "The following systems have been found to have at least one valid BIOS file.\n\n$systems_with_bios\n\nFor more information on the BIOS files found please use the Advanced check tool."
 
-  configurator_troubleshooting_tools_dialog
+  configurator_tools_and_troubleshooting_dialog
 }
 
 configurator_check_bios_files_advanced() {
@@ -678,10 +757,10 @@ configurator_check_bios_files_advanced() {
   --column "BIOS File Description" \
   "${bios_checked_list[@]}"
 
-  configurator_troubleshooting_tools_dialog
+  configurator_tools_and_troubleshooting_dialog
 }
 
-configurator_troubleshooting_tools_dialog() {
+configurator_tools_and_troubleshooting_dialog() {
   choice=$(zenity --list --title="RetroDECK Configurator Utility - Change Options" --cancel-label="Back" \
   --window-icon="/app/share/icons/hicolor/scalable/apps/net.retrodeck.retrodeck.svg" --width=1200 --height=720 \
   --column="Choice" --column="Action" \
@@ -1017,7 +1096,7 @@ configurator_welcome_dialog() {
   ;;
 
   "Tools and Troubleshooting" )
-    configurator_troubleshooting_tools_dialog
+    configurator_tools_and_troubleshooting_dialog
   ;;
 
   "Reset" )
